Vectorizing streets
The importer can now vectorize streets. To do so, the 'Adress' field is
mapped to 3 values. If the street is of the format
'street1 / street2'
the two streets will be mapped to ids which will live in the first 2
values, and the third becomes -1
If the street designation is of the form
'xxxx block of street1'
the first value becomes the street id, the second one the block number and
the third is +1.
Thus, the thrid parameter allows a kernel to destinguish between the two
types.